py-oracle-transfer/pyproject.toml
[tool.poetry]
name = "py-oracle-transfer"
version = "0.1.0"
description = ""
authors = ["Nicolas M. <mine.nicolas@gmail.com>"]
readme = "README.md"
packages = [{include = "py_oracle_transfer"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
pyyaml = "^6.0.1"
oracledb = "^2.3.0"
sqlalchemy = "^2.0.31"
click = "^8.1.7"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
